separe preconditions and precondition
